Progrress
- Added db service
- added auto Dependency injext for all services
- add data to table for 1st half of the request

+++ Backend/estate-emporium/Models/GetPropertyResponseModel.cs
@@ -9,7 +9,7 @@ namespace estate_emporium.Models
         /// </summary>
         /// <example>3</example>
         [Required]
-        public float price { get; set; }
+        public long price { get; set; }
         /// <summary>
         /// property ID
         /// </summary>
